Add files via upload
Build a stress polygon diagram to constrain the possible stress magnitudes for data from a well. 
Image logs from which the fractures are picked show no signs of drilling induced tensile fractures or wellbore breakouts. 
Use the key observations of the lack of wellbore failure combined with the relations for borehole stresses and frictional faulting theory to construct a stress polygon.